size_t/ptrdiff_t transition part 2: SFIO, hash lib and print(1)
This is the second of the thickfold patch series, which enables ksh93
to operate within a 64-bit address space.

The parts of ksh93 affected by this commit are:
- SFIO and the associated stdio wrapper. The sfvprintf and sfvscanf
  functions are notable for bearing the brunt of the substantial code
  changes.
  - Nearly all of the code is de novo, though towards the latter end
    of development I used graphviz commits for cross-reference:
    - https://gitlab.com/graphviz/graphviz/-/commit/6451a669
    - https://gitlab.com/graphviz/graphviz/-/commit/153a8f87
    - Improved multibyte handling in sfvscanf via use of
      unsigned char*; ported from graphviz:
      https://gitlab.com/graphviz/graphviz/-/commit/cb9b35d1
      (I'm aware this function is unused, but after some manual
      testing it works well AFAICT. A bit odd that ksh93 never
      uses it once.)
  - sfseek(): Removed a wasteful double assignment for f->iosz.
    This error was introduced in 2003-06-21 ksh93o+, and likely
    would have gone unnoticed if not for the thickfold project.
    (In practice this line was virtually always optimized out by the
    compiler, so it was dead code.)
- 64-bit modernization for the AST hash library.
- Transitioned pathgetlink() to return ssize_t.
- Updated the code in print.c for compatibility with the SFIO 64-bit
  modernization.
- Transitioned the strgrpmatch() family to use regflags_t to appease
  some -Wsign-conversion warnings (the type is now located in ast.h).
  - The now unused STR_INT flag backported from ksh93v- has been
    removed to avoid bitrot.
- Removed the unused nonstandard ssizeof() macro from ast.h.
- Made the snprintf and vsnprintf wrappers standards compliant
  with the C standard.

Change in the number of warnings on Linux when compiling with clang using
-Wsign-compare -Wshorten-64-to-32 -Wsign-conversion -Wimplicit-int-conversion:
3,808 => 3,291 => 248 (progression from part 1 => part 2 => part 13)

Progresses https://github.com/ksh93/ksh/issues/592

+++ src/lib/libast/include/sfio_t.h
@@ -76,7 +76,7 @@
 	  (short)(file),				/* file		*/ \
 	  (unsigned char*)(data),			/* data		*/ \
 	  (ssize_t)(size),				/* size		*/ \
-	  (ssize_t)(-1),				/* val		*/ \
+	  -1,						/* val		*/ \
 	  0,						/* extent	*/ \
 	  0,						/* here		*/ \
 	  0,						/* ngetr	*/ \
